+++ src/app/[variants]/(desktop)/desktop-onboarding/_layout/index.tsx
@@ -5,7 +5,8 @@ import { Divider } from 'antd';
 import { cx } from 'antd-style';
 import type { FC, PropsWithChildren } from 'react';
 
-import { SimpleTitleBar, TITLE_BAR_HEIGHT } from '@/features/ElectronTitlebar';
+import SimpleTitleBar from '@/features/Electron/titlebar/SimpleTitleBar';
+import { TITLE_BAR_HEIGHT } from '@/features/Electron/titlebar/constants';
 import LangButton from '@/features/User/UserPanel/LangButton';
 import ThemeButton from '@/features/User/UserPanel/ThemeButton';
 import { useIsDark } from '@/hooks/useIsDark';
