+++ src/app/[variants]/(main)/_layout/index.tsx
@@ -12,7 +12,8 @@ import Loading from '@/components/Loading/BrandTextLoading';
 import { isDesktop } from '@/const/version';
 import { BANNER_HEIGHT } from '@/features/AlertBanner/CloudBanner';
 import DesktopNavigationBridge from '@/features/DesktopNavigationBridge';
-import TitleBar, { TITLE_BAR_HEIGHT } from '@/features/ElectronTitlebar';
+import TitleBar from '@/features/Electron/titlebar/TitleBar';
+import { TITLE_BAR_HEIGHT } from '@/features/Electron/titlebar/constants';
 import HotkeyHelperPanel from '@/features/HotkeyHelperPanel';
 import NavPanel from '@/features/NavPanel';
 import { useFeedbackModal } from '@/hooks/useFeedbackModal';
