+++ src/features/Electron/titlebar/WinControl.tsx
@@ -1,0 +1,5 @@
+const WinControl = () => {
+  return <div style={{ width: 132 }} />;
+};
+
+export default WinControl;
